Implemented ModelViewSet

+++ apps/invite/views.py
@@ -1,4 +1,3 @@
-from rest_framework.generics import CreateAPIView, ListAPIView, UpdateAPIView, DestroyAPIView
 from rest_framework.permissions import AllowAny
 from rest_framework.response import Response
 from rest_framework.status import HTTP_200_OK, HTTP_400_BAD_REQUEST
@@ -11,35 +10,42 @@ from apps.invite import (
 )
 from apps.pokerboard import models as pokerboard_models
 
-
-class PokerboardMembersApi(ListAPIView):
-    serializer_class = invite_serializers.PokerboardMemberSerializer
-
-    def get(self, request, pk=None):
-        pokerboard = pokerboard_models.Pokerboard.objects.get(id=pk)
-        queryset = invite_models.Invite.objects.filter(pokerboard=pokerboard.id).filter(is_accepted=True)
-        pokerboard_member_serializer = self.serializer_class(queryset, many=True)
-        return Response(pokerboard_member_serializer.data)
-        
-
-class InviteUserApi(CreateAPIView):
-    """
-    API for inviting user to pokerboard
-    """
+class PokerboardMembersApi(ModelViewSet):
     serializer_class = invite_serializers.InviteUserSerializer
-    permission_classes = [invite_permissions.IsManagerPermission]
-
-
-class AcceptInviteApi(UpdateAPIView):
-    """
-    Invitation acception API
-    """
-    permission_classes = [AllowAny]
+    pokerboard_member_serializer = invite_serializers.PokerboardMemberSerializer
     queryset = invite_models.Invite.objects.all()
 
-    def put(self, request, pk = None, *args, **kwargs):
+    def get_permissions(self):
+        """
+        Instantiates and returns the list of permissions that this view requires.
+        """
+        permission_classes = [AllowAny]
+        if self.action == 'create' or self.action == 'destroy': 
+            permission_classes = [invite_permissions.IsManagerPermission]
+        return [permission() for permission in permission_classes]
+
+    def perform_create(self, serializer):
+        """
+        Invites a user/group to pokerboard
+        Only pokerboard's manager can perform this action
+        """
+        pokerboard = pokerboard_models.Pokerboard.objects.get(id=self.request.data['pokerboard'])
+        self.check_object_permissions(self.request, pokerboard)
+        return super().perform_create(serializer)
+
+    def retrieve(self, request, pk = None):
+        """
+        Gets all the pokerboard's members
+        """
+        invitee = invite_models.Invite.objects.filter(pokerboard=pk, is_accepted=True)
+        members = self.pokerboard_member_serializer(invitee, many=True)
+        return Response(members.data)
+
+    def update(self, request, pk = None, *args, **kwargs):
+        """
+        Invitation acception API
+        """
         instance = self.queryset.get(id=pk)
-        # TODO: Check if user's account is verified
         if instance.invitee == self.request.user.email and instance.is_accepted == False:
             instance.is_accepted = True
             instance.save()
@@ -47,11 +53,13 @@ class AcceptInviteApi(UpdateAPIView):
 
         return Response(status=HTTP_400_BAD_REQUEST)
 
-class RemoveInviteeApi(DestroyAPIView):
-    permission_classes = [invite_permissions.IsManagerPermission]
-    queryset = invite_models.Invite.objects.filter(is_accepted=True)
-
-    def delete(self, request, pk = None, *args, **kwargs):
+    def destroy(self, request, pk = None, *args, **kwargs):
+        """
+        Removes member from pokerboard
+        Only pokerboard's manager can perform this action
+        """
+        pokerboard = pokerboard_models.Pokerboard.objects.get(id=self.request.data['pokerboard'])
+        self.check_object_permissions(self.request, pokerboard)
         instance = self.queryset.get(id=pk)
         if instance:
             instance.delete()
